+++ src/OpenTelemetry.Api/Internal/ExceptionExtensions.cs
@@ -14,11 +14,7 @@
 // limitations under the License.
 // </copyright>
 
-#pragma warning disable IDE0005 // Temporarily suppressing "Using directive is unnecessary" until other projects have been updated. See #3958.
-using System;
 using System.Globalization;
-using System.Threading;
-#pragma warning restore IDE0005
 
 namespace OpenTelemetry.Internal
 {