+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/SdkLimitOptions.cs
@@ -21,10 +21,15 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation;
 
 internal sealed class SdkLimitOptions
 {
+    private const int DefaultSdkLimit = 128;
     private int? spanAttributeValueLengthLimit;
+    private bool spanAttributeValueLengthLimitSet;
     private int? spanAttributeCountLimit;
+    private bool spanAttributeCountLimitSet;
     private int? spanEventAttributeCountLimit;
+    private bool spanEventAttributeCountLimitSet;
     private int? spanLinkAttributeCountLimit;
+    private bool spanLinkAttributeCountLimitSet;
 
     internal SdkLimitOptions()
         : this(new ConfigurationBuilder().AddEnvironmentVariables().Build())
@@ -34,16 +39,16 @@ internal sealed class SdkLimitOptions
     internal SdkLimitOptions(IConfiguration configuration)
     {
         // https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/sdk-environment-variables.md#attribute-limits
-        SetIntConfigValue(configuration, "OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT", value => this.AttributeValueLengthLimit = value);
-        SetIntConfigValue(configuration, "OTEL_ATTRIBUTE_COUNT_LIMIT", value => this.AttributeCountLimit = value);
+        SetIntConfigValue(configuration, "OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT", value => this.AttributeValueLengthLimit = value, null);
+        SetIntConfigValue(configuration, "OTEL_ATTRIBUTE_COUNT_LIMIT", value => this.AttributeCountLimit = value, DefaultSdkLimit);
 
         // https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/sdk-environment-variables.md#span-limits
-        SetIntConfigValue(configuration, "OTEL_SPAN_ATTRIBUTE_VALUE_LENGTH_LIMIT", value => this.SpanAttributeValueLengthLimit = value);
-        SetIntConfigValue(configuration, "OTEL_SPAN_ATTRIBUTE_COUNT_LIMIT", value => this.SpanAttributeCountLimit = value);
-        SetIntConfigValue(configuration, "OTEL_SPAN_EVENT_COUNT_LIMIT", value => this.SpanEventCountLimit = value);
-        SetIntConfigValue(configuration, "OTEL_SPAN_LINK_COUNT_LIMIT", value => this.SpanLinkCountLimit = value);
-        SetIntConfigValue(configuration, "OTEL_EVENT_ATTRIBUTE_COUNT_LIMIT", value => this.SpanEventAttributeCountLimit = value);
-        SetIntConfigValue(configuration, "OTEL_LINK_ATTRIBUTE_COUNT_LIMIT", value => this.SpanLinkAttributeCountLimit = value);
+        SetIntConfigValue(configuration, "OTEL_SPAN_ATTRIBUTE_VALUE_LENGTH_LIMIT", value => this.SpanAttributeValueLengthLimit = value, null);
+        SetIntConfigValue(configuration, "OTEL_SPAN_ATTRIBUTE_COUNT_LIMIT", value => this.SpanAttributeCountLimit = value, null);
+        SetIntConfigValue(configuration, "OTEL_SPAN_EVENT_COUNT_LIMIT", value => this.SpanEventCountLimit = value, DefaultSdkLimit);
+        SetIntConfigValue(configuration, "OTEL_SPAN_LINK_COUNT_LIMIT", value => this.SpanLinkCountLimit = value, DefaultSdkLimit);
+        SetIntConfigValue(configuration, "OTEL_EVENT_ATTRIBUTE_COUNT_LIMIT", value => this.SpanEventAttributeCountLimit = value, null);
+        SetIntConfigValue(configuration, "OTEL_LINK_ATTRIBUTE_COUNT_LIMIT", value => this.SpanLinkAttributeCountLimit = value, null);
     }
 
     /// <summary>
@@ -64,8 +69,12 @@ internal sealed class SdkLimitOptions
     /// </remarks>
     public int? SpanAttributeValueLengthLimit
     {
-        get => this.spanAttributeValueLengthLimit ?? this.AttributeValueLengthLimit;
-        set => this.spanAttributeValueLengthLimit = value;
+        get => this.spanAttributeValueLengthLimitSet ? this.spanAttributeValueLengthLimit : this.AttributeValueLengthLimit;
+        set
+        {
+            this.spanAttributeValueLengthLimitSet = true;
+            this.spanAttributeValueLengthLimit = value;
+        }
     }
 
     /// <summary>
@@ -76,8 +85,12 @@ internal sealed class SdkLimitOptions
     /// </remarks>
     public int? SpanAttributeCountLimit
     {
-        get => this.spanAttributeCountLimit ?? this.AttributeCountLimit;
-        set => this.spanAttributeCountLimit = value;
+        get => this.spanAttributeCountLimitSet ? this.spanAttributeCountLimit : this.AttributeCountLimit;
+        set
+        {
+            this.spanAttributeCountLimitSet = true;
+            this.spanAttributeCountLimit = value;
+        }
     }
 
     /// <summary>
@@ -98,8 +111,12 @@ internal sealed class SdkLimitOptions
     /// </remarks>
     public int? SpanEventAttributeCountLimit
     {
-        get => this.spanEventAttributeCountLimit ?? this.SpanAttributeCountLimit;
-        set => this.spanEventAttributeCountLimit = value;
+        get => this.spanEventAttributeCountLimitSet ? this.spanEventAttributeCountLimit : this.SpanAttributeCountLimit;
+        set
+        {
+            this.spanEventAttributeCountLimitSet = true;
+            this.spanEventAttributeCountLimit = value;
+        }
     }
 
     /// <summary>
@@ -110,15 +127,23 @@ internal sealed class SdkLimitOptions
     /// </remarks>
     public int? SpanLinkAttributeCountLimit
     {
-        get => this.spanLinkAttributeCountLimit ?? this.SpanAttributeCountLimit;
-        set => this.spanLinkAttributeCountLimit = value;
+        get => this.spanLinkAttributeCountLimitSet ? this.spanLinkAttributeCountLimit : this.SpanAttributeCountLimit;
+        set
+        {
+            this.spanLinkAttributeCountLimitSet = true;
+            this.spanLinkAttributeCountLimit = value;
+        }
     }
 
-    private static void SetIntConfigValue(IConfiguration configuration, string key, Action<int> setter)
+    private static void SetIntConfigValue(IConfiguration configuration, string key, Action<int?> setter, int? defaultValue)
     {
         if (configuration.TryGetIntValue(key, out var result))
         {
             setter(result);
+        }
+        else if (defaultValue.HasValue)
+        {
+            setter(defaultValue);
         }
     }
 }