+++ src/OpenTelemetry/BatchExportProcessor.cs
@@ -16,10 +16,8 @@
 
 #nullable enable
 
-using System;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
-using System.Threading;
 using OpenTelemetry.Internal;
 
 namespace OpenTelemetry