+++ src/OpenTelemetry/Logs/LogRecordScope.cs
@@ -16,9 +16,7 @@
 
 #nullable enable
 
-using System;
 using System.Collections;
-using System.Collections.Generic;
 
 namespace OpenTelemetry.Logs
 {