+++ src/OpenTelemetry/Logs/OpenTelemetryLogger.cs
@@ -16,8 +16,6 @@
 
 #nullable enable
 
-using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Microsoft.Extensions.Logging;