+++ src/OpenTelemetry.Extensions.Hosting/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added the `OpenTelemetryBuilder.StartWithHost` extension.
+  ([#3923](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3923))
+
 ## 1.0.0-rc9.9
 
 Released 2022-Nov-07